# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ultra Venues API Docs",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/api/customer/register": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Register a new customer",
        "description": "This endpoint allows a new customer to register by providing their full name, email, phone number, and password.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Customer registration details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "fullName": {
                  "type": "string",
                  "example": "John Doe"
                },
                "email": {
                  "type": "string",
                  "example": "john.doe@example.com"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "1234567890"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              },
              "required": [
                "fullName",
                "email",
                "phoneNumber",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/customer/login": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Customer login",
        "description": "This endpoint allows a customer to login by providing their email and password.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Customer login details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "john.doe@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/customer/bookings/create": {
      "post": {
        "tags": [
          "Booking"
        ],
        "summary": "Create a new booking",
        "description": "This endpoint allows a customer to create a new booking by providing the necessary details.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Booking details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "eventType": {
                  "type": "string",
                  "example": "Wedding"
                },
                "startDate": {
                  "type": "string",
                  "example": "2024-07-20T00:00:00.000Z"
                },
                "endDate": {
                  "type": "string",
                  "example": "2024-07-21T00:00:00.000Z"
                },
                "customerEmailAddress": {
                  "type": "string",
                  "example": "john.doe@example.com"
                },
                "customerPhoneNumber": {
                  "type": "string",
                  "example": "1234567890"
                },
                "venuePhoneNumber": {
                  "type": "string",
                  "example": "0987654321"
                },
                "numberOfGuests": {
                  "type": "number",
                  "example": 150
                },
                "venueEmailAddress": {
                  "type": "string",
                  "example": "venue@example.com"
                },
                "price": {
                  "type": "number",
                  "example": 5000
                }
              },
              "required": [
                "eventType",
                "startDate",
                "endDate",
                "customerEmailAddress",
                "customerPhoneNumber",
                "venuePhoneNumber",
                "numberOfGuests",
                "venueEmailAddress",
                "price"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/customer/refresh-token": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Refresh access token",
        "description": "This endpoint allows a customer to refresh their access token using a refresh token.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Refresh token",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "type": "string",
                  "example": "your-refresh-token-here"
                }
              },
              "required": [
                "refreshToken"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}